+++ advents/src/main/kotlin/net/fish/geometry/paths/TrefoilPathCreator.kt
@@ -3,9 +3,9 @@ package net.fish.geometry.paths
 import net.fish.geometry.knots.Knots
 
 data class TrefoilPathCreator(
-    val scale: Double
+    val scale: Float
 ): PathCreator {
     override fun createPath(segments: Int): List<PathData> {
-        return Knots.wikiTrefoil(segments, scale)
+        return Knots.wikiTrefoil(segments, scale.toDouble())
     }
 }
